King's Own Tobacco

+++ code/game/objects/items/cigs.dm
@@ -419,6 +419,7 @@ LIGHTERS ARE IN LIGHTERS.DM
 /obj/item/clothing/mask/cigarette/rollie/Initialize(mapload)
 	. = ..()
 	scatter_atom()
+
 /obj/item/clothing/mask/cigarette/rollie/custom
 	desc = "A roll of dried plant matter wrapped in thin paper."
 	list_reagents = list()
@@ -478,20 +479,66 @@ LIGHTERS ARE IN LIGHTERS.DM
 		to_chat(user, "<span class='warning'>You need to dry this first!</span>")
 		return ITEM_INTERACT_COMPLETE
 
-	user.unequip(plant, TRUE)
 	user.unequip(src, TRUE)
 	var/obj/item/clothing/mask/cigarette/rollie/custom/custom_rollie = new (get_turf(user))
-	custom_rollie.reagents.maximum_volume = plant.reagents.total_volume
-	plant.reagents.trans_to(custom_rollie, plant.reagents.total_volume)
-	custom_rollie.smoketime = custom_rollie.reagents.total_volume * 2.5
-
-	user.put_in_active_hand(custom_rollie)
-	to_chat(user, "<span class='notice'>You roll the [plant.name] into a rolling paper.</span>")
-	custom_rollie.desc = "Dried [plant.name] rolled up in a thin piece of paper."
-
-	qdel(plant)
+	// Don't stuff the entire tin into a single cig, you barbarian!
+	if(istype(used, /obj/item/food/grown/tobacco/pre_dried))
+		custom_rollie.desc = "Dried tobacco rolled up in a thin piece of paper."
+		if(plant.reagents.total_volume > custom_rollie.chem_volume)
+			to_chat(user, "<span class='notice'>You pour some of [plant] into a rolling paper.</span>")
+			plant.reagents.trans_to(custom_rollie, 60)
+			plant.update_icon_state()
+		else
+			to_chat(user, "<span class='notice'>You empty [plant] into a rolling paper.</span>")
+			plant.reagents.trans_to(custom_rollie, plant.reagents.total_volume)
+			user.unequip(plant, TRUE)
+			qdel(plant)
+			user.put_in_active_hand(new /obj/item/trash/tobacco_tin)
+	else
+		user.unequip(plant, TRUE)
+		custom_rollie.reagents.maximum_volume = plant.reagents.total_volume
+		plant.reagents.trans_to(custom_rollie, plant.reagents.total_volume)
+		to_chat(user, "<span class='notice'>You roll the [plant.name] into a rolling paper.</span>")
+		custom_rollie.desc = "Dried [plant.name] rolled up in a thin piece of paper."
+		qdel(plant)
+	custom_rollie.smoketime = custom_rollie.reagents.total_volume * REAGENT_TIME_RATIO
+	user.put_in_hands(custom_rollie)
 	qdel(src)
 	return ITEM_INTERACT_COMPLETE
+
+/obj/item/food/grown/tobacco/pre_dried
+	name = "\improper King's Own tobacco"
+	desc = "A tin of loose-leaf tobacco for filling pipes and hand-rolled cigarettes."
+	icon = 'icons/obj/clothing/smoking.dmi'
+	icon_state = "pipe_tobacco"
+	seed = null
+	trash = /obj/item/trash/tobacco_tin
+	dry = TRUE
+	tastes = list("tobacco" = 1, "herbs" = 2, "spices" = 2)
+	volume = 240
+	list_reagents = list("nicotine" = 240) // One of these has to fill an entire pipe. This is also the same nicotine found in a pack of 6 normal cigarettes.
+
+/obj/item/food/grown/tobacco/pre_dried/examine_more(mob/user)
+	. = ..()
+	. += "A unique verity of hardy, arid tobacco grown on the Vulpkanin world of Strend in the Kelunian system. \
+	The strain was developed from varieties imported from Kelune by the first colonists, and is now considered a luxury brand."
+	. += ""
+	. += "Sun-dried, finely shredded, and infused with herbs and spices; it has a pleasant but difficult to pin down flavor.\
+	Whilst intended for use in pipes or hand-rolled cigarettes, some have been known to use it as dip or snuff."
+
+/obj/item/food/grown/tobacco/pre_dried/On_Consume(mob/M, mob/user)
+	update_icon_state()
+	..()
+
+/obj/item/food/grown/tobacco/pre_dried/update_icon_state()
+	. = ..()
+	if(reagents.total_volume < 240)
+		icon_state = "pipe_tobacco_open"
+
+/obj/item/trash/tobacco_tin
+	name = "empty tobacco tin"
+	icon = 'icons/obj/clothing/smoking.dmi'
+	icon_state = "pipe_tobacco_empty"
 
 //////////////////////////////
 // MARK: CIGARS
@@ -508,7 +555,7 @@ LIGHTERS ARE IN LIGHTERS.DM
 
 /obj/item/clothing/mask/cigarette/cigar/examine_more(mob/user)
 	. = ..()
-	. += "	Don't let the advertising fool you, this thing is a bargain basement, bottom-of-the-barrel product and the smoking experience it offers is little better than an oversized Robust cigarette."
+	. += "Don't let the advertising fool you, this thing is a bargain basement, bottom-of-the-barrel product and the smoking experience it offers is little better than an oversized Robust cigarette."
 	. += ""
 	. += "It still makes you look like a mafia boss, however."
 
@@ -667,10 +714,23 @@ LIGHTERS ARE IN LIGHTERS.DM
 		to_chat(user, "<span class='warning'>[src] is full!</span>")
 		return ITEM_INTERACT_COMPLETE
 
-	filler.reagents.trans_to(src, chem_volume)
-	to_chat(user, "<span class='notice'>You stuff the [filler.name] into the pipe.</span>")
+	if(istype(used, /obj/item/food/grown/tobacco/pre_dried))
+		to_chat(user, "<span class='notice'>You empty [filler] into the pipe.</span>")
+		if((chem_volume - reagents.total_volume) >= filler.reagents.total_volume)
+			filler.reagents.trans_to(src, chem_volume)
+			user.unequip(filler, TRUE)
+			qdel(filler)
+			user.put_in_active_hand(new /obj/item/trash/tobacco_tin)
+		else
+			to_chat(user, "<span class='notice'>You pour some of [filler] into the pipe.</span>")
+			filler.reagents.trans_to(src, clamp(filler.reagents.total_volume, 0, (chem_volume - reagents.total_volume)))
+			filler.update_icon_state()
+	else
+		to_chat(user, "<span class='notice'>You stuff the [filler.name] into the pipe.</span>")
+		filler.reagents.trans_to(src, chem_volume)
+		qdel(filler)
+
 	smoketime = max(reagents.total_volume * REAGENT_TIME_RATIO, smoketime)
-	qdel(filler)
 	return ITEM_INTERACT_COMPLETE
 
 /obj/item/clothing/mask/cigarette/pipe/light()
